refactor(profile): Renomeia e move fragmento de perfil
