+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -24,7 +24,7 @@
         tools:layout="@layout/fragment_notifications" />
     <fragment
         android:id="@+id/navigation_perfil"
-        android:name="com.mobile.app_iara.navigation_perfil"
-        android:label="fragment_navigation_perfil"
-        tools:layout="@layout/fragment_perfil" />
+        android:name="com.mobile.app_iara.ui.profile.Profile"
+        android:label="fragment_profile"
+        tools:layout="@layout/fragment_profile" />
 </navigation>
